// src/hooks/use-algorand/types.gen.ts
export type Account = {
	/**
	 * the account public key
	 */
	address: string;
	/**
	 * \[algo\] total number of MicroAlgos in the account
	 */
	amount: number;
	/**
	 * specifies the amount of MicroAlgos in the account, without the pending rewards.
	 */
	"amount-without-pending-rewards": number;
	/**
	 * \[appl\] applications local data stored in this account.
	 *
	 * Note the raw object uses `map[int] -> AppLocalState` for this type.
	 */
	"apps-local-state"?: Array<ApplicationLocalState>;
	/**
	 * \[teap\] the sum of all extra application program pages for this account.
	 */
	"apps-total-extra-pages"?: number;
	"apps-total-schema"?: ApplicationStateSchema;
	/**
	 * \[asset\] assets held by this account.
	 *
	 * Note the raw object uses `map[int] -> AssetHolding` for this type.
	 */
	assets?: Array<AssetHolding>;
	/**
	 * \[spend\] the address against which signing should be checked. If empty, the address of the current account is used. This field can be updated in any transaction by setting the RekeyTo field.
	 */
	"auth-addr"?: string;
	/**
	 * \[appp\] parameters of applications created by this account including app global data.
	 *
	 * Note: the raw account uses `map[int] -> AppParams` for this type.
	 */
	"created-apps"?: Array<Application>;
	/**
	 * \[apar\] parameters of assets created by this account.
	 *
	 * Note: the raw account uses `map[int] -> Asset` for this type.
	 */
	"created-assets"?: Array<Asset>;
	/**
	 * Whether or not the account can receive block incentives if its balance is in range at proposal time.
	 */
	"incentive-eligible"?: boolean;
	/**
	 * The round in which this account last went online, or explicitly renewed their online status.
	 */
	"last-heartbeat"?: number;
	/**
	 * The round in which this account last proposed the block.
	 */
	"last-proposed"?: number;
	/**
	 * MicroAlgo balance required by the account.
	 *
	 * The requirement grows based on asset and application usage.
	 */
	"min-balance": number;
	participation?: AccountParticipation;
	/**
	 * amount of MicroAlgos of pending rewards in this account.
	 */
	"pending-rewards": number;
	/**
	 * \[ebase\] used as part of the rewards computation. Only applicable to accounts which are participating.
	 */
	"reward-base"?: number;
	/**
	 * \[ern\] total rewards of MicroAlgos the account has received, including pending rewards.
	 */
	rewards: number;
	/**
	 * The round for which this information is relevant.
	 */
	round: number;
	/**
	 * Indicates what type of signature is used by this account, must be one of:
	 * * sig
	 * * msig
	 * * lsig
	 */
	"sig-type"?: "sig" | "msig" | "lsig";
	/**
	 * \[onl\] delegation status of the account's MicroAlgos
	 * * Offline - indicates that the associated account is delegated.
	 * *  Online  - indicates that the associated account used as part of the delegation pool.
	 * *   NotParticipating - indicates that the associated account is neither a delegator nor a delegate.
	 */
	status: string;
	/**
	 * The count of all applications that have been opted in, equivalent to the count of application local data (AppLocalState objects) stored in this account.
	 */
	"total-apps-opted-in": number;
	/**
	 * The count of all assets that have been opted in, equivalent to the count of AssetHolding objects held by this account.
	 */
	"total-assets-opted-in": number;
	/**
	 * \[tbxb\] The total number of bytes used by this account's app's box keys and values.
	 */
	"total-box-bytes"?: number;
	/**
	 * \[tbx\] The number of existing boxes created by this account's app.
	 */
	"total-boxes"?: number;
	/**
	 * The count of all apps (AppParams objects) created by this account.
	 */
	"total-created-apps": number;
	/**
	 * The count of all assets (AssetParams objects) created by this account.
	 */
	"total-created-assets": number;
};
// ...
export type AccountParticipation = {
	/**
	 * \[sel\] Selection public key (if any) currently registered for this round.
	 */
	"selection-participation-key": string;
	/**
	 * \[stprf\] Root of the state proof key (if any)
	 */
	"state-proof-key"?: string;
	/**
	 * \[voteFst\] First round for which this participation is valid.
	 */
	"vote-first-valid": number;
	/**
	 * \[voteKD\] Number of subkeys in each batch of participation keys.
	 */
	"vote-key-dilution": number;
	/**
	 * \[voteLst\] Last round for which this participation is valid.
	 */
	"vote-last-valid": number;
	/**
	 * \[vote\] root participation public key (if any) currently registered for this round.
	 */
	"vote-participation-key": string;
};
// ...
export type Application = {
	/**
	 * \[appidx\] application index.
	 */
	id: number;
	params: ApplicationParams;
};
// ...
export type ApplicationLocalState = {
	/**
	 * The application which this local state is for.
	 */
	id: number;
	"key-value"?: TealKeyValueStore;
	schema: ApplicationStateSchema;
};

/**
 * Stores the global information associated with an application.
 */
export type ApplicationParams = {
	/**
	 * \[approv\] approval program.
	 */
	"approval-program": string;
	/**
	 * \[clearp\] approval program.
	 */
	"clear-state-program": string;
	/**
	 * The address that created this application. This is the address where the parameters and global state for this application can be found.
	 */
	creator: string;
	/**
	 * \[epp\] the amount of extra program pages available to this app.
	 */
	"extra-program-pages"?: number;
	"global-state"?: TealKeyValueStore;
	"global-state-schema"?: ApplicationStateSchema;
	"local-state-schema"?: ApplicationStateSchema;
};
// ...
export type ApplicationStateSchema = {
	/**
	 * \[nbs\] num of byte slices.
	 */
	"num-byte-slice": number;
	/**
	 * \[nui\] num of uints.
	 */
	"num-uint": number;
};

/**
 * Specifies both the unique identifier and the parameters for an asset
 */
export type Asset = {
	/**
	 * unique asset identifier
	 */
	index: number;
	params: AssetParams;
};

/**
 * Describes an asset held by an account.
 *
 * Definition:
 * data/basics/userBalance.go : AssetHolding
 */
export type AssetHolding = {
	/**
	 * \[a\] number of units held.
	 */
	amount: number;
	/**
	 * Asset ID of the holding.
	 */
	"asset-id": number;
	/**
	 * \[f\] whether or not the holding is frozen.
	 */
	"is-frozen": boolean;
};
// ...
export type AssetParams = {
	/**
	 * \[c\] Address of account used to clawback holdings of this asset.  If empty, clawback is not permitted.
	 */
	clawback?: string;
	/**
	 * The address that created this asset. This is the address where the parameters for this asset can be found, and also the address where unwanted asset units can be sent in the worst case.
	 */
	creator: string;
	/**
	 * \[dc\] The number of digits to use after the decimal point when displaying this asset. If 0, the asset is not divisible. If 1, the base unit of the asset is in tenths. If 2, the base unit of the asset is in hundredths, and so on. This value must be between 0 and 19 (inclusive).
	 */
	decimals: number;
	/**
	 * \[df\] Whether holdings of this asset are frozen by default.
	 */
	"default-frozen"?: boolean;
	/**
	 * \[f\] Address of account used to freeze holdings of this asset.  If empty, freezing is not permitted.
	 */
	freeze?: string;
	/**
	 * \[m\] Address of account used to manage the keys of this asset and to destroy it.
	 */
	manager?: string;
	/**
	 * \[am\] A commitment to some unspecified asset metadata. The format of this metadata is up to the application.
	 */
	"metadata-hash"?: string;
	/**
	 * \[an\] Name of this asset, as supplied by the creator. Included only when the asset name is composed of printable utf-8 characters.
	 */
	name?: string;
	/**
	 * Base64 encoded name of this asset, as supplied by the creator.
	 */
	"name-b64"?: string;
	/**
	 * \[r\] Address of account holding reserve (non-minted) units of this asset.
	 */
	reserve?: string;
	/**
	 * \[t\] The total number of units of this asset.
	 */
	total: number;
	/**
	 * \[un\] Name of a unit of this asset, as supplied by the creator. Included only when the name of a unit of this asset is composed of printable utf-8 characters.
	 */
	"unit-name"?: string;
	/**
	 * Base64 encoded name of a unit of this asset, as supplied by the creator.
	 */
	"unit-name-b64"?: string;
	/**
	 * \[au\] URL where more information about the asset can be retrieved. Included only when the URL is composed of printable utf-8 characters.
	 */
	url?: string;
	/**
	 * Base64 encoded URL where more information about the asset can be retrieved.
	 */
	"url-b64"?: string;
};
// ...
export type TealKeyValueStore = Array<TealKeyValue>;